сделала правки по домашке и выполнила вторую часть
