+++ index.html
@@ -235,5 +235,6 @@
     </template>
     <script src="./js/main.js" type="module"></script>
     <script src="./js/functions.js"></script>
+    <script src="./js/util.js"></script>
 </body>
 </html>
