+++ js/main.js
@@ -1,1 +1,4 @@
 import './util.js';
+import {createBlockObjects} from './data.js';
+
+console.log(createBlockObjects(1, 25));
